requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.30.0
bcrypt==4.0.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
dotenv==0.9.9
email_validator==2.2.0
fastapi==0.115.12
greenlet==3.2.0
h11==0.14.0
idna==3.10
loguru==0.7.3
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
pydantic==2.11.3
pydantic_core==2.33.1
PyJWT==2.10.1
python-dotenv==1.1.0
python-multipart==0.0.20
python-slugify==8.0.4
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
text-unidecode==1.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.34.2
win32_setctime==1.2.0
